requirements.txt
requests>=2.32.4
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0
pydantic>=2.0.0
inquirer>=3.1.0
certifi>=2024.7.4
urllib3>=2.5.0
